[IMP] mail: chat window in mobile public page
With this commit the mobile view of the public page
is a fullscreen chat window just like regular discuss.

The sidebar on the mobile view is replaced by the
messaging menu just like regular discuss.

Added responsiveness to public welcome page.

task-3346088
task-2664822
