+++ addons/mail/static/src/core/common/chat_window.xml
@@ -67,6 +67,18 @@
                 <t t-set="action" t-value="threadActions.actions.at(-1)"/>
                 <t t-set="itemClass" t-value="'me-1'"/>
             </t>
+            <t t-if="this.store.inPublicPage and this.store.self.type === 'guest'">
+                <button class="btn ps-1" t-if="!state.editingGuestName">
+                    <img class="o-mail-Discuss-selfAvatar rounded-circle o_object_fit_cover flex-shrink-0" alt="Avatar" t-att-src="store.self.avatarUrl" t-on-click="() => state.editingGuestName = true"/>
+                </button>
+                <AutoresizeInput
+                    t-if="state.editingGuestName"
+                    className="'py-1 me-2'"
+                    autofocus="true"
+                    onValidate.bind="renameGuest"
+                    value="store.self.name"
+                />
+            </t>
         </div>
         <div t-if="!props.chatWindow.folded or ui.isSmall" class="bg-100 d-flex flex-column h-100 overflow-auto position-relative" t-ref="content">
             <t t-if="thread" name="thread content">
