+++ addons/mail/static/src/core/public_web/messaging_menu.scss
