+++ addons/mail/static/src/core/public_web/notification_item.scss
