+++ addons/mail/static/src/core/web/discuss_patch.js
@@ -1,7 +1,7 @@
 import { onRendered, useEffect } from "@odoo/owl";
 
 import { Discuss } from "@mail/core/public_web/discuss";
-import { MessagingMenu } from "@mail/core/web/messaging_menu";
+import { MessagingMenu } from "@mail/core/public_web/messaging_menu";
 
 import { ControlPanel } from "@web/search/control_panel/control_panel";
 import { _t } from "@web/core/l10n/translation";
