+++ addons/mail/static/src/core/web/discuss_patch.xml
@@ -4,9 +4,6 @@
         <xpath expr="//*[@t-ref='content']" position="before">
             <div t-if="ui.isSmall and store.discuss.activeTab === 'main'" class="w-100 border-bottom" t-call="mail.Discuss.mobileTopbar" t-ref="mobileTopbar"/>
         </xpath>
-        <xpath expr="//*[@t-ref='root']" position="inside">
-            <MessagingMenu t-if="ui.isSmall"/>
-        </xpath>
         <xpath expr="//*[@t-ref='root']" position="replace">
             <div class="h-100 d-flex flex-column">
                 <ControlPanel t-if="!ui.isSmall"/>
