+++ addons/mail/static/src/core/web/store_service_patch.js
@@ -1,6 +1,5 @@
 import { Record } from "@mail/core/common/record";
 import { Store } from "@mail/core/common/store_service";
-import { cleanTerm } from "@mail/utils/common/format";
 import { compareDatetime } from "@mail/utils/common/misc";
 import { browser } from "@web/core/browser/browser";
 import { _t } from "@web/core/l10n/translation";
@@ -25,83 +24,6 @@ const StorePatch = {
                 const getSortId = (activityGroup) =>
                     activityGroup.model === "mail.activity" ? Number.MAX_VALUE : activityGroup.id;
                 return getSortId(g1) - getSortId(g2);
-            },
-        });
-        this.menuThreads = Record.many("Thread", {
-            /** @this {import("models").Store} */
-            compute() {
-                /** @type {import("models").Thread[]} */
-                const searchTerm = cleanTerm(this.discuss.searchTerm);
-                let threads = Object.values(this.Thread.records).filter(
-                    (thread) =>
-                        (thread.displayToSelf ||
-                            (thread.needactionMessages.length > 0 &&
-                                thread.model !== "mail.box")) &&
-                        cleanTerm(thread.displayName).includes(searchTerm)
-                );
-                const tab = this.discuss.activeTab;
-                if (tab !== "main") {
-                    threads = threads.filter(({ channel_type }) =>
-                        this.tabToThreadType(tab).includes(channel_type)
-                    );
-                } else if (tab === "main" && this.env.inDiscussApp) {
-                    threads = threads.filter(({ channel_type }) =>
-                        this.tabToThreadType("mailbox").includes(channel_type)
-                    );
-                }
-                return threads;
-            },
-            /**
-             * @this {import("models").Store}
-             * @param {import("models").Thread} a
-             * @param {import("models").Thread} b
-             */
-            sort(a, b) {
-                /**
-                 * Ordering:
-                 * - threads with needaction
-                 * - unread channels
-                 * - read channels
-                 * - odoobot chat
-                 *
-                 * In each group, thread with most recent message comes first
-                 */
-                const aOdooBot = a.isCorrespondentOdooBot;
-                const bOdooBot = b.isCorrespondentOdooBot;
-                if (aOdooBot && !bOdooBot) {
-                    return 1;
-                }
-                if (bOdooBot && !aOdooBot) {
-                    return -1;
-                }
-                const aNeedaction = a.needactionMessages.length;
-                const bNeedaction = b.needactionMessages.length;
-                if (aNeedaction > 0 && bNeedaction === 0) {
-                    return -1;
-                }
-                if (bNeedaction > 0 && aNeedaction === 0) {
-                    return 1;
-                }
-                const aUnread = a.selfMember?.message_unread_counter;
-                const bUnread = b.selfMember?.message_unread_counter;
-                if (aUnread > 0 && bUnread === 0) {
-                    return -1;
-                }
-                if (bUnread > 0 && aUnread === 0) {
-                    return 1;
-                }
-                const aMessageDatetime = a.newestPersistentNotEmptyOfAllMessage?.datetime;
-                const bMessageDateTime = b.newestPersistentNotEmptyOfAllMessage?.datetime;
-                if (!aMessageDatetime && bMessageDateTime) {
-                    return 1;
-                }
-                if (!bMessageDateTime && aMessageDatetime) {
-                    return -1;
-                }
-                if (aMessageDatetime && bMessageDateTime && aMessageDatetime !== bMessageDateTime) {
-                    return bMessageDateTime - aMessageDatetime;
-                }
-                return b.localId > a.localId ? 1 : -1;
             },
         });
         this.inbox = Record.one("Thread");
