+++ addons/mail/static/src/discuss/call/web/messaging_menu_patch.js
@@ -1,4 +1,4 @@
-import { MessagingMenu } from "@mail/core/web/messaging_menu";
+import { MessagingMenu } from "@mail/core/public_web/messaging_menu";
 import { patch } from "@web/core/utils/patch";
 
 patch(MessagingMenu.prototype, {
