+++ addons/mail/static/src/discuss/core/public/welcome_page.xml
@@ -8,9 +8,9 @@
             <span t-else="">You've been invited to a chat!</span>
         </h1>
         <h2 class="m-5" t-esc="store.companyName"/>
-        <div class="d-flex justify-content-center">
-            <div t-if="this.store.discuss_public_thread.defaultDisplayMode === 'video_full_screen'" class="position-relative d-flex justify-content-center me-5" t-ref="root">
-                <video class="shadow rounded bg-dark" height="480" width="640" autoplay="" t-ref="video"/>
+        <div class="d-flex justify-content-center gap-5" t-att-class="{'flex-column': ui.isSmall}">
+            <div t-if="this.store.discuss_public_thread.defaultDisplayMode === 'video_full_screen'" class="position-relative d-flex justify-content-center" t-ref="root">
+                <video class="shadow rounded bg-dark" t-attf-height="{{ui.isSmall ? 240 : 480}}" t-attf-width="{{ui.isSmall ? 320 : 640}}" autoplay="" t-ref="video"/>
                 <p t-if="hasRtcSupport and !state.videoStream" class="position-absolute bottom-50 text-light">
                     Camera is off
                 </p>
@@ -19,10 +19,10 @@
                 </p>
                 <div class="position-absolute bottom-0">
                     <button class="btn fa-stack align-self-end p-0 m-3 rounded-circle fs-1 shadow" t-attf-class="{{ state.audioStream ? 'btn-dark' : 'btn-danger' }}" t-on-click="onClickMic">
-                        <i class="fa fa-stack" t-attf-class="{{ state.audioStream ? 'fa-microphone-slash' : 'fa-microphone' }}"/>
+                        <i class="fa" t-attf-class="{{ state.audioStream ? 'fa-microphone-slash' : 'fa-microphone' }}"/>
                     </button>
                     <button class="btn fa-stack align-self-end p-0 m-3 rounded-circle fs-1 shadow" t-attf-class="{{ state.videoStream ? 'btn-dark' : 'btn-danger' }}" t-on-click="onClickVideo">
-                        <i class="fa fa-camera fa-stack"/>
+                        <i class="fa fa-camera"/>
                     </button>
                 </div>
                 <audio autoplay="" t-ref="audio"/>
