+++ addons/rating/static/src/notification_item_patch.js
@@ -1,5 +1,5 @@
 /** @odoo-module */
 
-import { NotificationItem } from "@mail/core/web/notification_item";
+import { NotificationItem } from "@mail/core/public_web/notification_item";
 
 NotificationItem.props = [...NotificationItem.props, "rating?"];
