+++ addons/sms/static/src/messaging_menu/messaging_menu_patch.js
@@ -1,6 +1,6 @@
 /** @odoo-module */
 
-import { MessagingMenu } from "@mail/core/web/messaging_menu";
+import { MessagingMenu } from "@mail/core/public_web/messaging_menu";
 import { _t } from "@web/core/l10n/translation";
 import { patch } from "@web/core/utils/patch";
